implemented the wrapper, removed the old db. Wont work
